+++ lib/hycop/utils/hycop_utils.dart
@@ -3,13 +3,16 @@
 import 'package:flutter/material.dart';
 import 'package:appwrite/appwrite.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+
 import 'package:uuid/uuid.dart';
 import 'package:crypto/crypto.dart';
 import 'dart:convert';
 //import '../../common/util/config.dart';
 // import '../../common/util/logger.dart';
+import '../../common/util/config.dart';
 import '../enum/model_enums.dart';
 //import 'abs_database.dart';
+import '../hycop_factory.dart';
 import 'hycop_exceptions.dart';
 //import '../hycop_factory.dart';
 
@@ -128,4 +131,6 @@ class HycopUtils {
     }
     return uuid.replaceAll(RegExp(r'-'), '');
   }
+
+  static String get order => (HycopFactory.serverType == ServerType.supabase) ? 'order_' : 'order';
 }